update testing file
